% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery_site_generator.R
\name{gallery_site}
\alias{gallery_site}
\alias{gallery_site_generator}
\title{Gallery website generator}
\usage{
gallery_site(input, ...)

gallery_site_generator(input, ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory).}

\item{...}{Currently unused.}
}
\description{
Define a custom website generator to be used with \code{\link[rmarkdown:render_site]{rmarkdown::render_site()}}.
This generates a simple R Markdown website including a gallery of pages with
embedded content, based on metadata in JSON or YAML format and custom site
configuration options.
}
\details{
See \code{\link[rmarkdown:default_site_generator]{rmarkdown::default_site_generator()}}, from which the
implementation was adapted.
}
