% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery_site_generator.R
\name{gallery_site_generator}
\alias{gallery_site_generator}
\title{Gallery website generator.}
\usage{
gallery_site_generator(input, ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory).}

\item{...}{Currently unused.}
}
\description{
Gallery website generator.
}
\details{
See \code{\link[rmarkdown:default_site_generator]{rmarkdown::default_site_generator()}}, from which the
implementation was adapted.
}
