% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{gallery_content}
\alias{gallery_content}
\title{Gallery page conent}
\usage{
gallery_content(..., before = NULL, after = NULL)
}
\arguments{
\item{...}{Unnamed items included in the main \verb{<div>}}

\item{before, after}{The content of the \verb{<div>} before and after
the main. If character, the value is wrapped inside \code{\link[htmltools:HTML]{htmltools::HTML()}}.}
}
\value{
An HTML tag object that can be rendered as HTML using
  \code{\link{as.character}()}.
}
\description{
Create the HTML of a gallery page, comprising three \verb{<div>} elements: one
initial wrapping around \code{before}, one main comprising the arbitrary
\code{htmltools::tagList(...)}, and one final wrapping around \code{after}.
}
